Move out TypeTest to its own test file
